src/requirements.txt
click==8.1.2
Flask==2.0.3
Flask-CAS==1.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.1.1
mysql-connector==2.2.9
python-decouple==3.6
six==1.16.0
SQLAlchemy==1.4.32
uWSGI==2.0.20
Werkzeug==2.0.3
xmltodict==0.12.0
